Change tuple return to a class object for GetPaged()
